requirements.txt
flask
pillow
transformers
ultralytics
face-recognition
deepface
ollama
torch
torchvision
